% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_AFAD.R
\name{calc.Havershine}
\alias{calc.Havershine}
\title{Calculates Havershine Distances of the earthquake events to a given point's coordinates.}
\usage{
calc.Havershine(x = NULL, y = NULL, eqdata = NULL, R = NULL)
}
\arguments{
\item{x}{longitude of the fix point.}

\item{y}{latitude of the fix point.}

\item{eqdata}{Earthquake data downloaded as data frame from AFAD's \emph{Event Web Service}.}

\item{R}{Radius of the Earth; default is 6371 km.}
}
\value{
A vector of Havershine distances of earthquake events to the given coordiantes of a point.
}
\description{
Calculates Havershine Distances of downloads earthquake data from AFAD's \emph{Event Web Service} to a given point's coordinates as longitude and latitude.
}
\examples{
equake<-read.AFAD(eventid=539822)
opensmap.AFAD(equake,gap=.5,minnumtiles=20)
d<-calc.Havershine(35.3161,36.9637,equake,R=6378)
}
