% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats.R
\name{calc.meanCenter}
\alias{calc.meanCenter}
\title{Calculates Mean and Weighted Mean Center}
\usage{
calc.meanCenter(eqdata = NULL, type = NULL, weight = NULL)
}
\arguments{
\item{eqdata}{Downloaded earthquake data as data frame from AFAD's \emph{Event Web Service}.}

\item{type}{Type of the mean center as \strong{mean} for mean center and \strong{weighted} for weighted mean center. Default is "mean".}

\item{weight}{Weight data for weighted center mean.}
}
\value{
A data frame of earthquake events mean center consists of mean longitude and latitude.
}
\description{
Calculates mean and weighted mean centers of the downloaded earthquake data.
}
\examples{
equake<-read.AFAD(start="2022-01-01T12:00:00",end="2023-01-02T12:00:00",lat=35.5,lon=37,maxrad=format(100000,scientific=FALSE))
c.mean<-calc.meanCenter(equake,type="mean")
c.wmean<-calc.meanCenter(equake,type="weighted",equake$depth)
opensmap.AFAD(equake,gap=.5,minnumtiles=20)+geom_point(color="red",aes(x=c.mean[,1],y=c.mean[,2]))+geom_point(shape=8,color="red",aes(x=c.wmean[,1],y=c.wmean[,2]))

}
