% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats.R
\name{calc.StDevEllipse}
\alias{calc.StDevEllipse}
\title{Calculates Standard Deviation Ellipse.}
\usage{
calc.StDevEllipse(eqdata = NULL)
}
\arguments{
\item{eqdata}{Downloaded earthquake data as data frame from AFAD's \emph{Event Web Service}.}
}
\value{
A data frame containing related results.

\describe{
\item{Tetha}{Angle of the ellipse}
\item{sigma.x}{Deviation along the x axis}
\item{sigma.y}{Deviation along the y axis}
\item{area.ellipse}{Area of the std. deviation ellipse}
\item{A}{Major axis}
\item{B}{Minor axis}
\item{rotation}{Rotation correction}
}
}
\description{
Calculates standard deviation ellipse of the downloaded earthquake data.
}
\examples{
library(ggforce)
equake<-read.AFAD(start="2022-01-01T12:00:00",end="2023-01-02T12:00:00",lat=37.8,lon=38.3,maxrad=99999)
c.mean<-calc.meanCenter(equake,type="mean")
stdis<-calc.StDistance(equake)
stdev<-calc.StDevEllipse(equake)
p<-opensmap.AFAD(equake,gap=.5,minnumtiles=10)
p<-p+geom_point(color="red",aes(x=c.mean[,1],y=c.mean[,2]))
p<-p+geom_circle(aes(x0=c.mean[,1], y0=c.mean[,2], r=calc.StDistance(equake)), inherit.aes=FALSE)
p<-p+geom_ellipse(color="red",aes(x0=c.mean[,1],y0=c.mean[,2],angle=stdev$rot,a=stdev$A,b=stdev$B))
p

}
