% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats.R
\name{calc.StDistance}
\alias{calc.StDistance}
\title{Calculates Standard Distance}
\usage{
calc.StDistance(eqdata = NULL)
}
\arguments{
\item{eqdata}{Downloaded earthquake data as data frame from AFAD's \emph{Event Web Service}.}
}
\value{
Standard distance value.
}
\description{
Calculates standard distance of the downloaded earthquake data.
}
\examples{
equake<-read.AFAD(start="2022-01-01T12:00:00",end="2023-01-02T12:00:00",lat=37.8,lon=38.3,maxrad=99999)
c.mean<-calc.meanCenter(equake,type="mean")
stdis<-calc.StDistance(equake)
p<-opensmap.AFAD(equake,gap=.5,minnumtiles=10)
p+geom_point(color="red",aes(x=c.mean[,1],y=c.mean[,2]))
p+geom_circle(aes(x0=c.mean[,1], y0=c.mean[,2], r=calc.StDistance(equake)), inherit.aes=FALSE)

}
