% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{histog.magni}
\alias{histog.magni}
\title{Quick histogram of magnitude of downloaded earthquake data}
\usage{
histog.magni(
  eqdata,
  bins = NULL,
  col = NULL,
  fill = NULL,
  title = NULL,
  caption = NULL,
  xlabel = NULL,
  ylabel = NULL
)
}
\arguments{
\item{eqdata}{Downloaded earthquake data as data frame from AFAD's \emph{Event Web Service}.}

\item{bins}{Number of bins for histogram, default is 30.}

\item{col}{Color of plotting line.}

\item{fill}{Color of fill.}

\item{title}{The text for the title.}

\item{caption}{The text for the caption.}

\item{xlabel}{Title for x axis.}

\item{ylabel}{Title for y axis.}
}
\description{
Draws a quick ggplot2 histogram of \strong{magnitude} distribution of the downloaded earthquake data with a couple of plot settings from \emph{Event Web Service} API provided by AFAD.
}
\examples{
equake=read.AFAD(start = "2023-08-01",end="2023-08-31",lat=27.142826,lon=38.423733,maxrad=100000)
histog.magni(equake,col="red",fill = "black",title="August 2023 Eartquakes - Radial Search")
histog.depth(equake,col="red",fill = "black")
}
