% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_AFAD.R
\name{map.AFAD}
\alias{map.AFAD}
\title{Mapping of earthquake data using maps.}
\usage{
map.AFAD(eqdata = NULL, mapfile = NULL, ...)
}
\arguments{
\item{eqdata}{Earthquake data as data frame from AFAD's \emph{Event Web Service}.}

\item{mapfile}{Map file for graphical output.}

\item{...}{Other arguments to be passed to \emph{points}.}
}
\description{
Plots some simple charts described in the service depending on package \strong{maps} of the earthquake data from Event Web Service API provided by AFAD (Disaster and Emergency Management Authority of Republic of Turkiye).
}
\examples{
equake=read.AFAD(start = "2023-08-01",end="2023-08-15",lat=27.142826,lon=38.423733,maxrad=format(100000,scientific=FALSE))
map.AFAD(equake,"gadm36_TUR_2_sp.rds",col=equake$magnitude)
}
