% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_AFAD.R
\docType{data}
\name{read.AFAD}
\alias{read.AFAD}
\title{Read earthquake data from AFAD Event Web Service}
\usage{
read.AFAD
}
\arguments{
\item{start, end}{Time constraints limiting events occurring on or after the specified start time to on or before the specified end time. These constraints are required for all filters except \emph{eventid}.}

\item{minlat, maxlat, minlon, maxlon}{Southern, northern, western and eastern boundaries of the box search in \emph{degrees}, respectively. At least one of these constraints should be filled. If there is at least one empty constraint, the filter will declare boundaries and execute filtration.Box search is incompatible with radial search.}

\item{lat, lon, maxrad, minrad}{Specifies the latitude and longitude of a geographic point as a center point in \emph{degrees}. Using maximum and minimum distances from this geographic point in \emph{meters} defines a \emph{radial} or a \emph{circular} search according to minimum distance, respectively. These constraints should be given by user, except \emph{\strong{minrad}}.}

\item{minmag, maxmag, magtype}{Limit to events from a magnitude larger than or equal to the specified minimum to a magnitude smaller than or equal to the specified maximum in \emph{float} by type of magnitude in \emph{text} used to test minimum and maximum limits, respectively.}

\item{mindepth, maxdepth}{Limits to events from depths equal to or greater than the specified depth to depths less than or equal to the specified depth in \emph{km}, respectively.}

\item{orderby}{Valid orderby argument includes \emph{\emph{time}} order by origin ascending time as \emph{default}, \emph{\emph{timedesc}} order by origin descending time, \emph{\emph{magnitude}} order by ascending magnitude, \emph{\emph{magnitudedesc}} order by descending magnitude in \emph{text}.}

\item{eventid}{A unique ID number assigned to an event by the database. eventid suppresses all other arguments.}
}
\value{
A data frame of earthquake events with 15 variables provided by the API.
\describe{
\item{country}{Country which the earthquake event occured in the case of listing eathquake events in neighbour countries and/or seas.}
\item{date}{Date of the earthquake event occured.}
\item{depth}{Depth of the earthquake event occured in km.}
\item{district}{District which the earthquake event occured inside Turkiye's borders, too.}
\item{eventid}{Event ID assigned uniqly to the earthquake event.}
\item{latitude}{Latitude of the earthquake event.}
\item{location}{Exact location of the earthquake event.}
\item{longitude}{Longitude of the earthquake event.}
\item{magnitude}{Magnitude of the earthquake event.}
\item{neighborhood}{Neighborhood of the earthquake event.}
\item{province}{Province which the earthquake event.}
\item{rmd}{RMS of the earthquake event.}
\item{type}{Magnitude type of the earthquake event.}
\item{iseventupdate}{State of the earthquake event record as original record or an event record's update.}
\item{lastupdatedate}{Date of the last update of the earthquake event record.}
}
}
\description{
Downloads earthquake data from AFAD (Disaster and Emergency Management Authority of Republic of Turkiye) using \emph{Earthquake Web Service} API provided by AFAD.
}
\examples{
#An example for a specific earthquake data with an event id,
equake<-read.AFAD(eventid=512359)
#Another example for earthquake data with time and magnitute type constraints,
equake<-read.AFAD(start="2022-01-01T12:00:00",end="2023-01-02T12:00:00",magtype="Mw")
}
\keyword{datasets}
