% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turkiye.cities}
\alias{turkiye.cities}
\title{Turkish Cities' Longitude and Latitude Data}
\format{
A data frame with 81 rows and 3 columns:
\describe{
\item{City}{City's name}
\item{Latitude}{City's latitude}
\item{Longitude}{City's longitude}
}
}
\usage{
turkiye.cities
}
\description{
A small dataset containing of longitude and latitude data of 81 cities of Turkiye.
}
\details{
A small dataset containing of longitude and latitude data of 81 cities of Turkiye.
}
\references{
\url{http://www.beycan.net/1057/illerin-enlem-ve-boylamlari.html}
}
\keyword{dataset}
