% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turkiye_and_neighborhood_gazetteer}
\alias{turkiye_and_neighborhood_gazetteer}
\title{Turkiye and Neighborhood Gazetteer}
\format{
A data frame with 53125 rows and 6 columns:
\describe{
\item{FID}{The OID of the feature}
\item{NAME}{Approved name of the feature}
\item{LONGITUDE}{Geographic coordinate's longitude of the feature}
\item{LATITUDE}{Geographic coordinate's latitude of the feature}
\item{TYPE}{Geographic type of the feature}
\item{JOG.SHEET}{JOG sheet number in which the feature resides.}
}
}
\usage{
turkiye_and_neighborhood_gazetteer
}
\description{
A gazetteer represents the locations and proper names for map features around the Turkey and neighborhood.
}
\details{
This gazetteer provides a base map layer that may be used to find locations by their proper name anywhere around Turkey.
The categories include airports, coastal features, drainage features, land features, islands, sea features, political features, and populated places.
The coordinates of the bounding box from west to east bounding longitudes and from north to south bounding latitudes	are 25, 45, 43 and 35, respectively.
}
\references{
\url{https://www.harita.gov.tr/urunler/indirilebilir-verilerdosyalar/13}
}
\keyword{dataset}
